fix handle object

+++ pkg/vm/engine/disttae/logtailreplay/partition_state.go
@@ -454,6 +454,9 @@ func (p *PartitionState) HandleObjectInsert(
 		objEntry.Sorted = sortedCol[idx]
 
 		old, exist := p.dataObjects.Get(objEntry)
+		if exist {
+			objEntry.HasDeltaLoc = old.HasDeltaLoc
+		}
 		if exist && !old.IsEmpty() {
 			// why check the deleteTime here? consider this situation:
 			// 		1. insert on an object, then these insert operations recorded into a CKP.
@@ -485,9 +488,6 @@ func (p *PartitionState) HandleObjectInsert(
 				// only update object stats
 			}
 		} else {
-			if exist {
-				objEntry.HasDeltaLoc = old.HasDeltaLoc
-			}
 			e := ObjectIndexByTSEntry{
 				Time:         createTSCol[idx],
 				ShortObjName: *objEntry.ObjectShortName(),
@@ -518,16 +518,21 @@ func (p *PartitionState) HandleObjectInsert(
 			}
 		}
 
+		if objEntry.EntryState && objEntry.DeleteTime.IsEmpty() {
+			panic("logic error")
+		}
 		// for appendable object, gc rows when delete object
-		if objEntry.EntryState && !objEntry.DeleteTime.IsEmpty() {
-			iter := p.rows.Copy().Iter()
-			objID := objEntry.ObjectStats.ObjectName().ObjectId()
-			blkID := objectio.NewBlockidWithObjectID(objID, 0)
+		iter := p.rows.Copy().Iter()
+		objID := objEntry.ObjectStats.ObjectName().ObjectId()
+		trunctPoint := startTSCol[idx]
+		blkCnt := objEntry.ObjectStats.BlkCnt()
+		for i := uint32(0); i < blkCnt; i++ {
+
+			blkID := objectio.NewBlockidWithObjectID(objID, uint16(i))
 			pivot := RowEntry{
 				// aobj has only one blk
 				BlockID: *blkID,
 			}
-			trunctPoint := startTSCol[idx]
 			for ok := iter.Seek(pivot); ok; ok = iter.Next() {
 				entry := iter.Item()
 				if entry.BlockID != *blkID {
@@ -554,11 +559,25 @@ func (p *PartitionState) HandleObjectInsert(
 						blockDeleted++
 					}
 				}
+
+				//it's tricky here.
+				//Due to consuming lazily the checkpoint,
+				//we have to take the following scenario into account:
+				//1. CN receives deletes for a non-appendable block from the log tail,
+				//   then apply the deletes into PartitionState.rows.
+				//2. CN receives block meta of the above non-appendable block to be inserted
+				//   from the checkpoint, then apply the block meta into PartitionState.blocks.
+				// So , if the above scenario happens, we need to set the non-appendable block into
+				// PartitionState.dirtyBlocks.
+				if !objEntry.EntryState && !objEntry.HasDeltaLoc {
+					p.dirtyBlocks.Set(entry.BlockID)
+					break
+				}
 			}
 			iter.Release()
 
 			// if there are no rows for the block, delete the block from the dirty
-			if scanCnt == blockDeleted && p.dirtyBlocks.Len() > 0 {
+			if objEntry.EntryState && scanCnt == blockDeleted && p.dirtyBlocks.Len() > 0 {
 				p.dirtyBlocks.Delete(*blkID)
 			}
 		}
